refactor(frontend): move hooks inside AnswerForm
